+++ backend/src/services/InscripcionService.js
@@ -164,46 +164,3 @@ export const validarInscripcion = async ({ tutorId, inscripcionId, acepta }) => 
     return { mensaje: 'Respuesta registrada correctamente.' };
 };
 
-export const obtenerSolicitudesPendientes = async (tutorUsuarioId) => {
-    const tutor = await prisma.tutor.findUnique({
-        where: { usuario_id: tutorUsuarioId }
-    });
-
-    if (!tutor) {
-        throw new Error('No se encontró el tutor.');
-    }
-
-    const solicitudes = await prisma.inscripcion_tutor.findMany({
-        where: {
-            tutor_id: tutor.id,
-            aprobado: false
-        },
-        include: {
-            inscripcion: true,
-            competidor: {
-                include: {
-                    usuario: true,
-                    colegio: true,
-                    provincia: {
-                        include: {
-                            departamento: true
-                        }
-                    }
-                }
-            }
-        }
-    });
-
-    return solicitudes.map((s) => ({
-        id: s.id,
-        competidor: {
-            nombre: s.competidor.usuario.nombre,
-            apellido: s.competidor.usuario.apellido,
-            colegio: s.competidor.colegio.nombre_colegio,
-            provincia: s.competidor.provincia.nombre_provincia,
-            departamento: s.competidor.provincia.departamento.nombre_departamento
-        },
-        fecha_inscripcion: s.inscripcion.fecha_inscripcion,
-        estado: s.inscripcion.estado_inscripcion
-    }));
-};
